% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_expand.R
\name{copy}
\alias{copy}
\title{Copy an entire object}
\usage{
copy(x)
}
\arguments{
\item{x}{object}
}
\value{
copy of the object 'x'
}
\description{
Mainly intended to copy data.table objects because by default they are modified by reference. See example.
}
\examples{
data(mtcars)
dt_mtcars = as.data.table(mtcars)
dt_mtcars2 = dt_mtcars
dt_mtcars3 = copy(dt_mtcars)
let(dt_mtcars, new = 1)

head(dt_mtcars2) # we see 'new' column
head(dt_mtcars3) # no 'new' column
}
