% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_utils.R
\name{dt_count}
\alias{dt_count}
\title{Additional functions}
\usage{
dt_count(data, ..., weight = NULL, sort = FALSE, name = "n")
}
\arguments{
\item{data}{data.table/data.frame data.frame will be automatically converted
to data.table.}

\item{...}{variables to group by.}

\item{sort}{logical. If TRUE result will be sorted in desending order by resulting variable.}

\item{name}{character. Name of resulting variable.}

\item{weight.}{optional. Unquoted variable name. If provided result will be the sum of this variable by groups.}
}
\value{
data.table
}
\description{
Additional functions
}
\examples{
1
}
