% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/let.R
\name{take}
\alias{take}
\alias{let}
\alias{take_if}
\alias{let_if}
\title{Title}
\usage{
take(data, ..., by, keyby, with = TRUE,
  nomatch = getOption("datatable.nomatch"), mult = "all", roll = FALSE,
  rollends = if (roll == "nearest") c(TRUE, TRUE) else if (roll >= 0) c(FALSE,
  TRUE) else c(TRUE, FALSE), which = FALSE, .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"), drop = NULL,
  on = NULL, autoname = TRUE, fun = NULL)

let(data, ..., by, keyby, with = TRUE,
  nomatch = getOption("datatable.nomatch"), mult = "all", roll = FALSE,
  rollends = if (roll == "nearest") c(TRUE, TRUE) else if (roll >= 0) c(FALSE,
  TRUE) else c(TRUE, FALSE), which = FALSE, .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"), drop = NULL,
  on = NULL)

take_if(data, i, ..., by, keyby, with = TRUE,
  nomatch = getOption("datatable.nomatch"), mult = "all", roll = FALSE,
  rollends = if (roll == "nearest") c(TRUE, TRUE) else if (roll >= 0) c(FALSE,
  TRUE) else c(TRUE, FALSE), which = FALSE, .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"), drop = NULL,
  on = NULL, autoname = TRUE, fun = NULL)

let_if(data, i, ..., by, keyby, with = TRUE,
  nomatch = getOption("datatable.nomatch"), mult = "all", roll = FALSE,
  rollends = if (roll == "nearest") c(TRUE, TRUE) else if (roll >= 0) c(FALSE,
  TRUE) else c(TRUE, FALSE), which = FALSE, .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"), drop = NULL,
  on = NULL)
}
\arguments{
\item{data}{sa}

\item{...}{sa}

\item{by}{as}

\item{keyby}{as}

\item{with}{as}

\item{nomatch}{as}

\item{mult}{as}

\item{roll}{as}

\item{rollends}{as}

\item{which}{as}

\item{.SDcols}{as}

\item{verbose}{as}

\item{allow.cartesian}{as}

\item{drop}{as}

\item{on}{as}

\item{autoname}{as}

\item{fun}{as}

\item{i}{asa}
}
\value{
sa
}
\description{
Title
}
\examples{
1
}
