% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_wide.R
\name{to_long}
\alias{to_long}
\alias{to_wide}
\title{Convert data to long or to wide form}
\usage{
to_long(
  data,
  columns,
  names_column = "variable",
  values_column = "value",
  keep,
  drop_na = FALSE,
  names_factor = TRUE,
  value_factor = FALSE,
  ...
)

to_wide(
  data,
  names_column = variable,
  values_column = value,
  keep,
  fun = identity,
  sep = "_",
  fill = NA,
  drop = TRUE,
  ...
)
}
\arguments{
\item{data}{A data.frame object to convert}

\item{columns}{Variables for stacking. When missing, we will stack all columns outside \code{keep} columns.}

\item{names_column}{name of the stacked variable names column. The default name is 'variable'.}

\item{values_column}{name of the stacked data values column(s). The default
name is 'value'. Multiple names can be provided here for the case when
\code{columns} is a list, though note well that the names provided in
\code{columns} take precedence.}

\item{keep}{columns which will be kept as is, e. g. only recycled or
deduplicated. If missing, it is all columns except stacked or unstacked.}

\item{drop_na}{If TRUE, NA values will be removed from the stacked data.}

\item{names_factor}{If TRUE, the column with names will be converted to
factor, else it will be a character column. TRUE by default.}

\item{value_factor}{If TRUE, the value column will be converted to factor,
else the stacked values type is left unchanged. FALSE by default.}

\item{...}{other arguments passed to \code{data.table::melt}/\code{data.table::dcast}}

\item{fun}{Function.}

\item{sep}{A character.}

\item{fill}{= NA,}

\item{drop}{A logical. TRUE by default.}
}
\value{
data.table in the wide or long form.
}
\description{
Convert data to long or to wide form
}
\examples{
1
}
