% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maditr.R
\docType{package}
\name{maditr}
\alias{maditr}
\alias{maditr-package}
\title{maditr: Tables with Labels and Some Useful Functions from Spreadsheets and SPSS Statistics}
\description{
'expss' package implements some popular functions from spreadsheets and SPSS
Statistics software. Implementations are not complete copies of their
originals. I try to make them consistent with other R functions. See examples
in the vignette and in the help.
}
