% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query_if}
\title{Title}
\usage{
query(data, j, by, keyby, with = TRUE,
  nomatch = getOption("datatable.nomatch"), mult = "all", roll = FALSE,
  rollends = if (roll == "nearest") c(TRUE, TRUE) else if (roll >= 0) c(FALSE,
  TRUE) else c(TRUE, FALSE), which = FALSE, .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"), drop = NULL,
  on = NULL)

query_if(data, i, j, by, keyby, with = TRUE,
  nomatch = getOption("datatable.nomatch"), mult = "all", roll = FALSE,
  rollends = if (roll == "nearest") c(TRUE, TRUE) else if (roll >= 0) c(FALSE,
  TRUE) else c(TRUE, FALSE), which = FALSE, .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"), drop = NULL,
  on = NULL)
}
\arguments{
\item{data}{sa}

\item{by}{as}

\item{keyby}{as}

\item{with}{as}

\item{nomatch}{as}

\item{mult}{as}

\item{roll}{as}

\item{rollends}{as}

\item{which}{as}

\item{.SDcols}{as}

\item{verbose}{as}

\item{allow.cartesian}{as}

\item{drop}{as}

\item{on}{as}

\item{i}{asa}

\item{...}{sa}

\item{autoname}{as}

\item{fun}{as}
}
\value{
sa
}
\description{
Title
}
\examples{
1
}
