% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maditr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{\%<>\%}
\alias{reexports}
\alias{\%$\%}
\alias{reexports}
\alias{\%T>\%}
\alias{reexports}
\alias{data.table}
\alias{reexports}
\alias{tables}
\alias{reexports}
\alias{setkey}
\alias{reexports}
\alias{setkeyv}
\alias{reexports}
\alias{key}
\alias{reexports}
\alias{key<-}
\alias{reexports}
\alias{haskey}
\alias{reexports}
\alias{CJ}
\alias{reexports}
\alias{SJ}
\alias{reexports}
\alias{copy}
\alias{reexports}
\alias{set2key}
\alias{reexports}
\alias{set2keyv}
\alias{reexports}
\alias{key2}
\alias{reexports}
\alias{setindex}
\alias{reexports}
\alias{setindexv}
\alias{reexports}
\alias{indices}
\alias{reexports}
\alias{as.data.table}
\alias{reexports}
\alias{is.data.table}
\alias{reexports}
\alias{test.data.table}
\alias{reexports}
\alias{last}
\alias{reexports}
\alias{first}
\alias{reexports}
\alias{like}
\alias{reexports}
\alias{\%like\%}
\alias{reexports}
\alias{between}
\alias{reexports}
\alias{\%between\%}
\alias{reexports}
\alias{inrange}
\alias{reexports}
\alias{\%inrange\%}
\alias{reexports}
\alias{timetaken}
\alias{reexports}
\alias{truelength}
\alias{reexports}
\alias{alloc.col}
\alias{reexports}
\alias{:=}
\alias{reexports}
\alias{setattr}
\alias{reexports}
\alias{setnames}
\alias{reexports}
\alias{setcolorder}
\alias{reexports}
\alias{set}
\alias{reexports}
\alias{setDT}
\alias{reexports}
\alias{setDF}
\alias{reexports}
\alias{setorder}
\alias{reexports}
\alias{setorderv}
\alias{reexports}
\alias{setNumericRounding}
\alias{reexports}
\alias{getNumericRounding}
\alias{reexports}
\alias{chmatch}
\alias{reexports}
\alias{\%chin\%}
\alias{reexports}
\alias{chorder}
\alias{reexports}
\alias{chgroup}
\alias{reexports}
\alias{rbindlist}
\alias{reexports}
\alias{fread}
\alias{reexports}
\alias{fwrite}
\alias{reexports}
\alias{foverlaps}
\alias{reexports}
\alias{shift}
\alias{reexports}
\alias{transpose}
\alias{reexports}
\alias{tstrsplit}
\alias{reexports}
\alias{frank}
\alias{reexports}
\alias{frankv}
\alias{reexports}
\alias{address}
\alias{reexports}
\alias{.SD}
\alias{reexports}
\alias{.N}
\alias{reexports}
\alias{.I}
\alias{reexports}
\alias{.GRP}
\alias{reexports}
\alias{.BY}
\alias{reexports}
\alias{rleid}
\alias{reexports}
\alias{rleidv}
\alias{reexports}
\alias{rowid}
\alias{reexports}
\alias{rowidv}
\alias{reexports}
\alias{as.xts.data.table}
\alias{reexports}
\alias{uniqueN}
\alias{reexports}
\alias{setDTthreads}
\alias{reexports}
\alias{getDTthreads}
\alias{reexports}
\alias{fintersect}
\alias{reexports}
\alias{fsetdiff}
\alias{reexports}
\alias{funion}
\alias{reexports}
\alias{fsetequal}
\alias{reexports}
\alias{shouldPrint}
\alias{reexports}
\alias{fsort}
\alias{reexports}
\alias{groupingsets}
\alias{reexports}
\alias{cube}
\alias{reexports}
\alias{rollup}
\alias{reexports}
\alias{melt}
\alias{reexports}
\alias{dcast}
\alias{reexports}
\alias{as.IDate}
\alias{reexports}
\alias{as.ITime}
\alias{reexports}
\alias{IDateTime}
\alias{reexports}
\alias{second}
\alias{reexports}
\alias{minute}
\alias{reexports}
\alias{hour}
\alias{reexports}
\alias{yday}
\alias{reexports}
\alias{wday}
\alias{reexports}
\alias{mday}
\alias{reexports}
\alias{week}
\alias{reexports}
\alias{isoweek}
\alias{reexports}
\alias{month}
\alias{reexports}
\alias{quarter}
\alias{reexports}
\alias{year}
\alias{reexports}
\alias{as.chron.IDate}
\alias{reexports}
\alias{as.chron.ITime}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{data.table}}, \code{\link[data.table]{tables}}, \code{\link[data.table]{setkey}}, \code{\link[data.table]{setkeyv}}, \code{\link[data.table]{key}}, \code{\link[data.table]{key<-}}, \code{\link[data.table]{haskey}}, \code{\link[data.table]{CJ}}, \code{\link[data.table]{SJ}}, \code{\link[data.table]{copy}}, \code{\link[data.table]{set2key}}, \code{\link[data.table]{set2keyv}}, \code{\link[data.table]{key2}}, \code{\link[data.table]{setindex}}, \code{\link[data.table]{setindexv}}, \code{\link[data.table]{indices}}, \code{\link[data.table]{as.data.table}}, \code{\link[data.table]{is.data.table}}, \code{\link[data.table]{test.data.table}}, \code{\link[data.table]{last}}, \code{\link[data.table]{first}}, \code{\link[data.table]{like}}, \code{\link[data.table]{\%like\%}}, \code{\link[data.table]{between}}, \code{\link[data.table]{\%between\%}}, \code{\link[data.table]{inrange}}, \code{\link[data.table]{\%inrange\%}}, \code{\link[data.table]{timetaken}}, \code{\link[data.table]{truelength}}, \code{\link[data.table]{alloc.col}}, \code{\link[data.table]{:=}}, \code{\link[data.table]{setattr}}, \code{\link[data.table]{setnames}}, \code{\link[data.table]{setcolorder}}, \code{\link[data.table]{set}}, \code{\link[data.table]{setDT}}, \code{\link[data.table]{setDF}}, \code{\link[data.table]{setorder}}, \code{\link[data.table]{setorderv}}, \code{\link[data.table]{setNumericRounding}}, \code{\link[data.table]{getNumericRounding}}, \code{\link[data.table]{chmatch}}, \code{\link[data.table]{\%chin\%}}, \code{\link[data.table]{chorder}}, \code{\link[data.table]{chgroup}}, \code{\link[data.table]{rbindlist}}, \code{\link[data.table]{fread}}, \code{\link[data.table]{fwrite}}, \code{\link[data.table]{foverlaps}}, \code{\link[data.table]{shift}}, \code{\link[data.table]{transpose}}, \code{\link[data.table]{tstrsplit}}, \code{\link[data.table]{frank}}, \code{\link[data.table]{frankv}}, \code{\link[data.table]{address}}, \code{\link[data.table]{.SD}}, \code{\link[data.table]{.N}}, \code{\link[data.table]{.I}}, \code{\link[data.table]{.GRP}}, \code{\link[data.table]{.BY}}, \code{\link[data.table]{rleid}}, \code{\link[data.table]{rleidv}}, \code{\link[data.table]{rowid}}, \code{\link[data.table]{rowidv}}, \code{\link[data.table]{as.xts.data.table}}, \code{\link[data.table]{uniqueN}}, \code{\link[data.table]{setDTthreads}}, \code{\link[data.table]{getDTthreads}}, \code{\link[data.table]{fintersect}}, \code{\link[data.table]{fsetdiff}}, \code{\link[data.table]{funion}}, \code{\link[data.table]{fsetequal}}, \code{\link[data.table]{shouldPrint}}, \code{\link[data.table]{fsort}}, \code{\link[data.table]{groupingsets}}, \code{\link[data.table]{cube}}, \code{\link[data.table]{rollup}}, \code{\link[data.table]{melt}}, \code{\link[data.table]{dcast}}, \code{\link[data.table]{as.IDate}}, \code{\link[data.table]{as.ITime}}, \code{\link[data.table]{IDateTime}}, \code{\link[data.table]{second}}, \code{\link[data.table]{minute}}, \code{\link[data.table]{hour}}, \code{\link[data.table]{yday}}, \code{\link[data.table]{wday}}, \code{\link[data.table]{mday}}, \code{\link[data.table]{week}}, \code{\link[data.table]{isoweek}}, \code{\link[data.table]{month}}, \code{\link[data.table]{quarter}}, \code{\link[data.table]{year}}, \code{\link[data.table]{as.chron.IDate}}, \code{\link[data.table]{as.chron.ITime}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%$\%}}, \code{\link[magrittr]{\%T>\%}}}
}}

