% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_list.R
\name{to_list}
\alias{to_list}
\alias{to_vec}
\alias{to_df}
\alias{to_dfr}
\alias{to_dfc}
\title{Title}
\usage{
to_list(
  data,
  expr = NULL,
  ...,
  skip_null = TRUE,
  trace = FALSE,
  trace_step = 1L
)

to_vec(
  data,
  expr = NULL,
  ...,
  skip_null = TRUE,
  trace = FALSE,
  trace_step = 1L,
  recursive = TRUE,
  use.names = TRUE
)

to_df(data, expr = NULL, ..., trace = FALSE, trace_step = 1L, idcol = NULL)

to_dfr(data, expr = NULL, ..., trace = FALSE, trace_step = 1L, idcol = NULL)

to_dfc(data, expr = NULL, ..., trace = FALSE, trace_step = 1)
}
\arguments{
\item{data}{data.frame/list/vector}

\item{expr}{expression}

\item{skip_null}{logical Should we skip NULL's from result? Default is TRUE}
}
\value{
list
}
\description{
Title
}
\examples{
1
}
